Imported and Updated
Imported and placed your tint machine, I also added multiple footstep sounds. I also updated and added basic animations. Since this is going to be a single player game, I have made it very simple animations. I also increased the height of the player but kept camera height the same so the players head is not in the way when sprinting.

Next plans:
1) Make a music playlist that randomly selects songs

2) Create a UI for the computer

3) Create an intro UI

4) Start making NPCs walk in and out of the store. Maybe make a little bell sound as they walk in?

5) I would like to add a level system as we get a little further in development so there is some more of a directed progression. For example, you can't buy a certain tint until you're this level.

6) Add a save function

7) Make new desks/workstations

8) Start working on other level designs for store upgrades

9) Create a way for NPCs to order

10) Start working on coding all the possible tint combinations and percentages/amounts of tint.

11) I'm not sure what else I need to do after that