Fixed seating and Object Collision ~Kinda~
Seating still has one error message appearing, but it is very non-fatal whatsoever. There is one error with picking up object collision but has little to no effect at the moment. The radio dissapears when it is being held and has the "GrabbedObject" layer which disabled collision. I now just compromised to make it where items that can be picked up have no collision with the player. I plan to move on tomorrow from this step as I have spent WAY TOO MUCH time on this step of implementation. I hope to start on the shop UI soon