Fixed resolution bug & other
Fixed the bug with resolution where the build made the resolution zoomed way in. Also implemented the maxFps but have not added it. I made the radiant website able to click now, but have some extra things to work on. I also added a sensitivity option in settings but have not implemented it whatsoever. Game is coming along well.