Added UI Selection
Added the ability to select items in the UI. Will add ability to pour tint soon. Will need to implement the ability to stir as well after tint pouring. So far so good. I do need to work on NPCs however and their movements. Such as a queue and stop when they achieve their position.