Order Management
Created the ability to sell paint. Each can can be sold. There is a slight bug for mixing paint and I have not completed the entire selling thing, however, you can create orders using i and you gain score based on how close you are to the value +3 for being exact +2 for being within 3 +1 for being within 10. Press F when the can is on the white countertop to sell. I have not increased the balance when selling since I have not decided on economy yet.