Created buying system
Created a system to purchase paint cans. I have not set a display for the price or anything like that, but I do have each white can as $9.99 to start. There is a way to soft-lock yourself if you do shit for the first 3 cans, so I may add a loan website later down the line. This is an exciting update since the game is actually a playable game now!