Imported a few tools / Started work on save
Started work on save state controller, there is now a save/load system for the inventory. There is still a lot to be done, however, I have done enough work for the inventory and balance to create a save and load state. Using G as a save and H as a load state. I still have a lot to go for saving, but I so far so good. I did not use that save tool for creating this save state thing, I'm just going to do it from scratch. There is still a lot to do, but it is a good start. It isn't set up for use yet, but it saves an in-game state as of right now, does not save on exit or to a file as of right now.