Saving between launches
Implemented the ability to save and load between launches. There is still more data that needs grabbing as we develop more, but it is in a functional state. I will have to do more with this to load the correct world and load on startup, but there is now saving. (Press G to save to a file, and on restart press H to load from that file)