Manual Tinter
Manual Tinter added with all 3 animations. Had an issue importing it as an FBX file so they are .blend files. The main tinter file has the red to green anim, other two models are named accordingly and hold the other two animations. I confirmed that they work on the main model.